app/requirements.txt
openai
face_recognition
python-dotenv
opencv-python
dlib
pillow
matplotlib
selenium
